Feature/reddit

+++ assets/js/gallery.js
@@ -1,10 +1,18 @@
+// Get search form element
+const searchFormEl = document.getElementById('search-form');
+searchFormEl.addEventListener('submit', function (e) {
+  e.preventDefault();
+  const queryStr = document.getElementById('query-input').value;
+  getGallery(queryStr);
+});
+
 // Function to render to the DOM the results of an image query from the images.nasa.gov API
 
 function getGallery(queryStr) {
   const url = 'https://images-api.nasa.gov/search?media_type=image&q=';
 
   // if query submitted, use it, else just look for black holes
-  const query = queryStr || 'black hole';
+  const query = queryStr || 'mexico';
   const params = { params: { q: query, media_type: 'image' } };
 
   axios
@@ -37,9 +45,20 @@ function getGallery(queryStr) {
       // get copy of current localStorage object
       const localStorageObj = JSON.parse(localStorage.userInfo);
 
-      // push latest search overview to array then save updated localStorage object
-      localStorageObj.recent.push(recentSearchObj);
-      localStorage.setItem('userInfo', JSON.stringify(localStorageObj));
+      let alreadyRecented = false;
+      for (let i = 0; i < localStorageObj.recent.length; i++) {
+        if (localStorageObj.recent[i].query == recentSearchObj.query) {
+          alreadyRecented = true;
+        }
+      }
+
+      // if recent search is not in there already
+      if (!alreadyRecented) {
+        // push latest search overview to array then save updated localStorage object
+        localStorageObj.recent.push(recentSearchObj);
+        localStorage.setItem('userInfo', JSON.stringify(localStorageObj));
+        populateRecents();
+      }
 
       // TO DO - render the first 10 results at first, render the next 10 if user clicks "Load more..." or scrolls down, then the next 10, and so on... so that the browser isn't inundated with 100 image loads all at once (test this)
 
@@ -48,4 +67,30 @@ function getGallery(queryStr) {
     .catch((err) => console.error(err));
 }
 
+function populateRecents() {
+  // get copy of current localStorage object
+  const localStorageObj = JSON.parse(localStorage.userInfo);
+
+  clearSectionById('#recents-container');
+
+  // get recent section element
+  const recentSectionEl = document.getElementById('recents-container');
+
+  localStorageObj.recent.forEach((search) => {
+    const divEl = document.createElement('div');
+    divEl.classList =
+      'grid justify-items-center inline-block bg-center bg-indigo-100 rounded-lg p-2';
+    const h3El = document.createElement('h3');
+    h3El.textContent = '"' + search.query + '"';
+    const imgEl = document.createElement('img');
+    imgEl.src = search.thumbnail;
+    imgEl.alt = search.query;
+    imgEl.classList.add('rounded-lg');
+
+    divEl.append(h3El, imgEl);
+    recentSectionEl.append(divEl);
+  });
+}
+
 getGallery();
+populateRecents();
